// src/interfaces/models.ts
export interface IUser {
  id?: number | null;
  name?: string | null;
  email?: string | null;
  email_verified_at?: string | null;
  password?: string | null;
  phone_personal?: string | null;
  phone_professional?: string | null;
  url_photo?: string | null;
  status?: string | null;
  created_at?: string | null;
  created_by?: string | null;
  created?: IUser | null;
  updated_at?: string | null;
  updated_by?: string | null;
  updated?: IUser | null;
  deleted_at?: string | null;
  deleted_by?: string | null;
  deleted?: IUser | null;
}

export interface ISaasClient {
  id?: number | null;
  name?: string | null;
  email?: string | null;
  phone?: string | null;
  observation?: string | null;
  status?: string | null;
  created_at?: string | null;
  created_by?: string | null;
  created?: IUser | null;
  updated_at?: string | null;
  updated_by?: string | null;
  updated?: IUser | null;
  deleted_at?: string | null;
  deleted_by?: string | null;
  deleted?: IUser | null;
}

export interface IPermission {
  id?: number | null;
  name?: string | null;
  created_at?: string | null;
  created_by?: string | null;
  created?: IUser | null;
  updated_at?: string | null;
  updated_by?: string | null;
  updated?: IUser | null;
  deleted_at?: string | null;
  deleted_by?: string | null;
  deleted?: IUser | null;
}

export interface IEvent {
  id?: number | null;
  name?: string | null;
  description?: string | null;
  url_photo?: string | null;
  saas_client_id?: string | null;
  saas_client?: ISaasClient | null;
  created_at?: string | null;
  created_by?: number | null;
  created?: IUser | null;
  updated_at?: string | null;
  updated_by?: number | null;
  updated?: IUser | null;
  deleted_at?: string | null;
  deleted_by?: number | null;
  deleted?: IUser | null;
}

export interface IEventList {
  id?: number | null;
  event_id?: number | null;
  event?: IEvent | null;
  name?: string | null;
  description?: string | null;
  url_photo?: string | null;
  saas_client_id?: number | null;
  saas_client?: ISaasClient | null;
  created_at?: string | null;
  created_by?: number | null;
  created?: IUser | null;
  updated_at?: string | null;
  updated_by?: number | null;
  updated?: IUser | null;
  deleted_at?: string | null;
  deleted_by?: number | null;
  deleted?: IUser | null;
}

export interface IEventListItem {
  id?: number | null;
  name?: string | null;
  email?: string | null;
  phone?: string | null;
  event_id?: number | null;
  event?: IEvent | null;
  event_list_id?: number | null;
  event_list?: IEventList | null;
  payment_status?: string | null;
  saas_client_id?: number | null;
  saas_client?: ISaasClient | null;
  created_at?: string | null;
  created_by?: number | null;
  created?: IUser | null;
  updated_at?: string | null;
  updated_by?: number | null;
  updated?: IUser | null;
  deleted_at?: string | null;
  deleted_by?: number | null;
  deleted?: IUser | null;
}